Remove the (int) cast, since you can't take address of a value which has been casted.  Instead we change the underlying type in the structure to be an int.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@6834 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/asn.1/asn1_k_decode.c
@@ -663,7 +663,7 @@ asn1_error_code asn1_decode_etype_info_entry(buf, val)
   { begin_structure();
     get_field(val->etype,0,asn1_decode_ui_4);
     if (tagnum == 1) {
-	    get_lenfield((int) val->length,val->salt,1,asn1_decode_octetstring);
+	    get_lenfield(val->length,val->salt,1,asn1_decode_octetstring);
     } else {
 	    val->length = 0;
 	    val->salt = 0;
